# UniversityStats/UniversityStats.Client/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "UniversityStats.API",
    "version": "1.0"
  },
  "paths": {
    "/api/Department": {
      "get": {
        "tags": [
          "Department"
        ],
        "summary": "Return list of departments",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepartmentDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepartmentDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepartmentDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Department"
        ],
        "summary": "Add department to database",
        "requestBody": {
          "description": "Department's information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Department"
        ],
        "summary": "Correct department's informations",
        "requestBody": {
          "description": "Department's information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Department/{departmentId}": {
      "get": {
        "tags": [
          "Department"
        ],
        "summary": "Find department by department's id",
        "parameters": [
          {
            "name": "departmentId",
            "in": "path",
            "description": "Department's id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Department"
        ],
        "summary": "Delete department's information by department's id",
        "parameters": [
          {
            "name": "departmentId",
            "in": "path",
            "description": "Department's id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/DepartmentSpecialty": {
      "get": {
        "tags": [
          "DepartmentSpecialty"
        ],
        "summary": "Return list of (department specialty)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepartmentSpecialtyDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepartmentSpecialtyDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepartmentSpecialtyDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "DepartmentSpecialty"
        ],
        "summary": "Post (Department specialty) to database",
        "requestBody": {
          "description": "(Department specialty)'s information in format DTO",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentSpecialtyDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentSpecialtyDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentSpecialtyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentSpecialtyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentSpecialtyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentSpecialtyDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "DepartmentSpecialty"
        ],
        "summary": "Correct (department specialty)'s information by specialty's id",
        "requestBody": {
          "description": "(Department specialty)'s information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentSpecialtyDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentSpecialtyDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentSpecialtyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentSpecialtyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentSpecialtyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentSpecialtyDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/DepartmentSpecialty/{specialtyId}": {
      "get": {
        "tags": [
          "DepartmentSpecialty"
        ],
        "summary": "Find (department specialty)'s information by specialty's id",
        "parameters": [
          {
            "name": "specialtyId",
            "in": "path",
            "description": "Specialty's id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentSpecialtyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentSpecialtyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentSpecialtyDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "DepartmentSpecialty"
        ],
        "summary": "Delete (department specialty) by specialty's id",
        "parameters": [
          {
            "name": "specialtyId",
            "in": "path",
            "description": "Specialty's id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Specialty/{specialtyId}": {
      "get": {
        "tags": [
          "Specialty"
        ],
        "summary": "Return specialty's information by specialty's id",
        "parameters": [
          {
            "name": "specialtyId",
            "in": "path",
            "description": "Specialty's id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialtyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialtyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialtyDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Specialty"
        ],
        "summary": "Delete specialty by specialty's id",
        "parameters": [
          {
            "name": "specialtyId",
            "in": "path",
            "description": "Specialty's id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DepartmentDto": {
        "type": "object",
        "properties": {
          "facultyId": {
            "type": "string",
            "description": "Faculty's id",
            "nullable": true
          },
          "nameDepartment": {
            "type": "string",
            "description": "Department's name",
            "nullable": true
          },
          "departmentId": {
            "type": "string",
            "description": "Department's id",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Class data transfer about department's information"
      },
      "DepartmentSpecialtyDto": {
        "type": "object",
        "properties": {
          "departmentId": {
            "type": "string",
            "description": "Department's id",
            "nullable": true
          },
          "specialtyId": {
            "type": "string",
            "description": "Specialty's id",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Class for data transfer about information of department and specialty"
      },
...
      "SpecialtyDto": {
        "type": "object",
        "properties": {
          "specialtyId": {
            "type": "string",
            "description": "Specialty's id",
            "nullable": true
          },
          "nameSpecialty": {
            "type": "string",
            "description": "Specialty's name",
            "nullable": true
          },
          "numberOfGroups": {
            "type": "integer",
            "description": "Total groups in specialty",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Class for data transfer about specialty's information"
      },
...
    }
  }
}